+++ examples/layout.rs
@@ -1,4 +1,4 @@
-use enigo::{Enigo, Key, KeyboardControllable, MouseControllable};
+use enigo::{Enigo, Key, KeyboardControllable};
 use std::thread;
 use std::time::Duration;
 
@@ -6,16 +6,12 @@ fn main() {
     thread::sleep(Duration::from_secs(4));
     let mut enigo = Enigo::new();
 
-    println!("Pressing pagedown");
     enigo.key_click(Key::PageDown);
-
+    enigo.key_click(enigo::Key::UpArrow);
     enigo.key_click(enigo::Key::UpArrow);
     enigo.key_click(enigo::Key::DownArrow);
     enigo.key_click(enigo::Key::LeftArrow);
+    enigo.key_click(enigo::Key::LeftArrow);
     enigo.key_click(enigo::Key::RightArrow);
-    enigo.key_sequence("𝕊");
-
-    // Special chars
-    // Need two u16s to be encoded
-    // 𝕊
+    enigo.key_sequence("𝕊"); // Special char which needs two u16s to be encoded
 }